+++ modules/console/builtin/cpp/Gateway.cpp
@@ -11,17 +11,18 @@
 #include "clcBuiltin.hpp"
 #include "inputBuiltin.hpp"
 #include "terminal_sizeBuiltin.hpp"
+#include "consoleboxBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
 const std::wstring gatewayName = L"console";
 //=============================================================================
 static const nlsGateway gateway[] = {
+    { "consolebox", (ptrBuiltin)Nelson::ConsoleGateway::consoleboxBuiltin, -1, 1, CPP_BUILTIN },
     { "clc", (ptrBuiltin)Nelson::ConsoleGateway::clcBuiltin, 0, 0, CPP_BUILTIN_WITH_EVALUATOR },
     { "input", (ptrBuiltin)Nelson::ConsoleGateway::inputBuiltin, 1, 1, CPP_BUILTIN_WITH_EVALUATOR },
     { "terminal_size", (ptrBuiltin)Nelson::ConsoleGateway::terminal_sizeBuiltin, -1, 0,
         CPP_BUILTIN_WITH_EVALUATOR }
-
 };
 //=============================================================================
 NLSGATEWAYFUNC(gateway)