+++ modules/console/src/include/ConsoleBox.hpp
@@ -7,25 +7,19 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "MacroCompleter.hpp"
-#include "PathFunctionIndexerManager.hpp"
-#include "StringHelpers.hpp"
+#pragma once
+//=============================================================================
+#include "nlsConsole_exports.h"
 //=============================================================================
 namespace Nelson {
+NLSCONSOLE_IMPEXP bool
+ConsoleBoxShow(bool show);
 //=============================================================================
-wstringVector
-MacroCompleter(const std::wstring& prefix)
-{
-    wstringVector res;
-    wstringVector macros = PathFunctionIndexerManager::getInstance()->getMacrosList(false);
-    res.reserve(macros.size());
-    for (const auto& macro : macros) {
-        if (StringHelpers::starts_with(macro, prefix)) {
-            res.emplace_back(macro);
-        }
-    }
-    return res;
+NLSCONSOLE_IMPEXP bool
+ConsoleBoxToggle();
+//=============================================================================
+NLSCONSOLE_IMPEXP bool
+ConsoleBoxIsVisible();
+//=============================================================================
 }
-//=============================================================================
-} // namespace Nelson;
 //=============================================================================